# libs/local/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "OpenWhisk REST API",
        "description": "API for OpenWhisk",
        "version": "0.1.0"
    },
    "produces": [
        "application/json"
    ],
    "basePath": "/api/v1",
...
    "paths": {
...
        "/namespaces/{namespace}/actions/{actionName}": {
            "get": {
                "tags": [
                    "Actions"
                ],
                "summary": "Get action information",
                "description": "Get action information.",
                "operationId": "getActionByName",
                "parameters": [
                    {
                        "name": "namespace",
                        "in": "path",
                        "description": "The entity namespace",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "actionName",
                        "in": "path",
                        "description": "Name of action to fetch",
                        "required": true,
                        "type": "string"
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Returned action",
                        "schema": {
                            "$ref": "#/definitions/Action"
                        }
                    },
                    "401": {
                        "$ref": "#/responses/UnauthorizedRequest"
                    },
                    "404": {
                        "$ref": "#/responses/ItemNotFound"
                    },
                    "500": {
                        "$ref": "#/responses/ServerError"
                    }
                }
            },
            "put": {
                "tags": [
                    "Actions"
                ],
                "description": "Create or update an action",
                "summary": "Create or update an action",
                "operationId": "updateAction",
                "parameters": [
                    {
                        "name": "namespace",
                        "in": "path",
                        "description": "The entity namespace",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "actionName",
                        "in": "path",
                        "description": "Name of action",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "overwrite",
                        "in": "query",
                        "description": "Overwrite item if it exists. Default is false.",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "true",
                            "false"
                        ]
                    },
                    {
                        "name": "action",
                        "in": "body",
                        "description": "The action being updated",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ActionPut"
                        }
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "$ref": "#/responses/UpdatedItem"
                    },
                    "400": {
                        "$ref": "#/responses/BadRequest"
                    },
                    "401": {
                        "$ref": "#/responses/UnauthorizedRequest"
                    },
                    "409": {
                        "$ref": "#/responses/Conflict"
                    },
                    "413": {
                        "$ref": "#/responses/RequestEntityTooLarge"
                    },
                    "500": {
                        "$ref": "#/responses/ServerError"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Actions"
                ],
                "description": "Delete an action",
                "summary": "Delete an action",
                "operationId": "deleteAction",
                "parameters": [
                    {
                        "name": "namespace",
                        "in": "path",
                        "description": "The entity namespace",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "actionName",
                        "in": "path",
                        "description": "Name of action",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "$ref": "#/responses/DeletedItem"
                    },
                    "401": {
                        "$ref": "#/responses/UnauthorizedRequest"
                    },
                    "404": {
                        "$ref": "#/responses/ItemNotFound"
                    },
                    "500": {
                        "$ref": "#/responses/ServerError"
                    }
                }
            },
            "post": {
                "tags": [
                    "Actions"
                ],
                "description": "Invoke an action",
                "summary": "Invoke an action",
                "operationId": "invokeAction",
                "parameters": [
                    {
                        "name": "namespace",
                        "in": "path",
                        "description": "The entity namespace",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "actionName",
                        "in": "path",
                        "description": "Name of action",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "blocking",
                        "in": "query",
                        "description": "Blocking or non-blocking invocation. Default is non-blocking.",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "true",
                            "false"
                        ]
                    },
                    {
                        "name": "result",
                        "in": "query",
                        "description": "Return only the result of a blocking activation. Default is false.",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "true",
                            "false"
                        ]
                    },
                    {
                        "name": "timeout",
                        "in": "query",
                        "description": "Wait no more than specified duration in milliseconds for a blocking response. Default value and max allowed timeout are 60000.",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "payload",
                        "in": "body",
                        "description": "The parameters for the action being invoked",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/KeyValue"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful activation",
                        "schema": {
                            "$ref": "#/definitions/Activation"
                        }
                    },
                    "202": {
                        "description": "Activation request accepted",
                        "$ref": "#/responses/AcceptedActivation"
                    },
                    "401": {
                        "$ref": "#/responses/UnauthorizedRequest"
                    },
                    "404": {
                        "$ref": "#/responses/ItemNotFound"
                    },
                    "408": {
                        "$ref": "#/responses/Timeout"
                    },
                    "500": {
                        "$ref": "#/responses/ServerError"
                    },
                    "502": {
                        "description": "Activation produced an application error",
                        "schema": {
                            "$ref": "#/definitions/Activation"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "KeyValue": {
            "required": [
                "key",
                "value"
            ],
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "description": "Any JSON value"
                }
            }
        },
        "ItemId": {
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        },
        "ErrorMessage": {
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "ActionLimits": {
            "description": "Limits on a specific action",
            "properties": {
                "timeout": {
                    "type": "integer",
                    "format": "int32",
                    "default": 30000
                },
                "memory": {
                    "type": "integer",
                    "format": "int32",
                    "default": 256
                }
            }
        },
...
        "Action": {
            "required": [
                "namespace",
                "name",
                "version",
                "publish",
                "exec",
                "parameters",
                "limits"
            ],
            "properties": {
                "namespace": {
                    "type": "string",
                    "description": "Namespace of the item",
                    "minLength": 1
                },
                "name": {
                    "type": "string",
                    "description": "Name of the item",
                    "minLength": 1
                },
                "version": {
                    "type": "string",
                    "description": "Semantic version of the item",
                    "minLength": 1
                },
                "publish": {
                    "type": "boolean",
                    "description": "Whether to publish the item or not"
                },
                "exec": {
                    "$ref": "#/definitions/ActionExec"
                },
                "annotations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/KeyValue"
                    },
                    "description": "annotations on the item"
                },
                "parameters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/KeyValue"
                    },
                    "description": "parameter bindings included in the context passed to the action"
                },
                "limits": {
                    "$ref": "#/definitions/ActionLimits"
                }
            }
        },
        "ActionPut": {
            "description": "A restricted Action view that elides properties that are auto-assigned or derived from the URI (i.e., the namespace and name).",
            "properties": {
                "version": {
                    "type": "string",
                    "description": "Semantic version of the item",
                    "minLength": 1
                },
                "publish": {
                    "type": "boolean",
                    "description": "Whether to publish the item or not"
                },
                "exec": {
                    "$ref": "#/definitions/ActionExec"
                },
                "annotations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/KeyValue"
                    },
                    "description": "annotations on the item"
                },
                "parameters": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/KeyValue"
                    },
                    "description": "parameter bindings included in the context passed to the action"
                },
                "limits": {
                    "$ref": "#/definitions/ActionLimits"
                }
            }
        },
        "ActionExec": {
            "required": [
                "kind"
            ],
            "properties": {
                "kind": {
                    "type": "string",
                    "enum": [
                        "nodejs",
                        "nodejs:6",
                        "nodejs:default",
                        "python",
                        "swift",
                        "swift:3",
                        "swift:default",
                        "java",
                        "blackbox"
                    ],
                    "description": "the type of action"
                },
                "code": {
                    "type": "string",
                    "description": "The code to execute when kind is not 'blackbox'"
                },
                "image": {
                    "type": "string",
                    "description": "container image name when kind is 'blackbox'"
                },
                "init": {
                    "type": "string",
                    "description": "optional zipfile reference when code kind is 'nodejs'"
                }
            },
            "description": "definition of the action, such as javascript code or the name of a container"
        },
...
        "Activation": {
            "required": [
                "namespace",
                "name",
                "version",
                "publish",
                "subject",
                "activationId",
                "start",
                "end",
                "result",
                "logs"
            ],
            "properties": {
                "namespace": {
                    "type": "string",
                    "description": "Namespace of the associated item"
                },
                "name": {
                    "type": "string",
                    "description": "Name of the item"
                },
                "version": {
                    "type": "string",
                    "description": "Semantic version of the item"
                },
                "publish": {
                    "type": "boolean",
                    "description": "Whether to publish the item or not"
                },
                "subject": {
                    "type": "string",
                    "description": "The subject that activated the item"
                },
                "activationId": {
                    "type": "string",
                    "description": "Id of the activation"
                },
                "start": {
                    "type": "string",
                    "description": "Time when the activation began"
                },
                "end": {
                    "type": "string",
                    "description": "Time when the activation completed"
                },
                "result": {
                    "$ref": "#/definitions/ActivationResult"
                },
                "logs": {
                    "type": "string",
                    "description": "Logs generated by the activation"
                }
            }
        },
...
        "ActivationResult": {
            "properties": {
                "status": {
                    "type": "string",
                    "description": "Exit status of the activation"
                },
                "value": {
                    "description": "The return value from the activation"
                }
            }
        },
...
    },
    "responses": {
        "BadRequest": {
            "description": "Bad request",
            "schema": {
                "$ref": "#/definitions/ErrorMessage"
            }
        },
        "UnauthorizedRequest": {
            "description": "Unauthorized request",
            "schema": {
                "$ref": "#/definitions/ErrorMessage"
            }
        },
        "ServerError": {
            "description": "Server error",
            "schema": {
                "$ref": "#/definitions/ErrorMessage"
            }
        },
...
        "UpdatedItem": {
            "description": "Updated Item",
            "schema": {
                "$ref": "#/definitions/ItemId"
            }
        },
        "DeletedItem": {
            "description": "Deleted Item"
        },
        "ItemNotFound": {
            "description": "Item not found",
            "schema": {
                "$ref": "#/definitions/ErrorMessage"
            }
        },
        "Timeout": {
            "description": "Request timed out"
        },
        "Conflict": {
            "description": "Conflicting item already exists",
            "schema": {
                "$ref": "#/definitions/ErrorMessage"
            }
        },
        "AcceptedActivation": {
            "description": "Accepted activation request",
            "schema": {
                "$ref": "#/definitions/ItemId"
            }
        },
...
        "RequestEntityTooLarge": {
            "description": "Request entity too large",
            "schema": {
                "$ref": "#/definitions/ErrorMessage"
            }
        }
    }
}